+++ src/At.ts
@@ -1,0 +1,32 @@
+/**
+ * @since 2.3.0
+ */
+import * as I from './internal'
+import { Iso } from './Iso'
+import { Lens } from './Lens'
+
+// -------------------------------------------------------------------------------------
+// model
+// -------------------------------------------------------------------------------------
+
+/**
+ * @category model
+ * @since 2.3.0
+ */
+export interface At<S, I, A> {
+  readonly at: (i: I) => Lens<S, A>
+}
+
+// -------------------------------------------------------------------------------------
+// constructors
+// -------------------------------------------------------------------------------------
+
+/**
+ * Lift an instance of `At` using an `Iso`
+ *
+ * @category constructors
+ * @since 2.3.0
+ */
+export const fromIso = <T, S>(iso: Iso<T, S>) => <I, A>(sia: At<S, I, A>): At<T, I, A> => ({
+  at: (i) => I.isoComposeLens(sia.at(i))(iso)
+})