+++ src/At/Set.ts
@@ -1,7 +1,7 @@
 /**
  * @since 1.2.0
  */
-import { At, Lens } from '../index'
+import { At, Lens } from '..'
 import { Eq } from 'fp-ts/lib/Eq'
 import * as S from 'fp-ts/lib/Set'