+++ src/Index/Array.ts
@@ -1,7 +1,7 @@
 /**
  * @since 1.2.0
  */
-import { Index, Optional } from '../index'
+import { Index, Optional } from '..'
 import { lookup, updateAt } from 'fp-ts/lib/Array'
 import { isNone } from 'fp-ts/lib/Option'