+++ src/Traversal.ts
@@ -1,0 +1,135 @@
+/**
+ * @since 2.3.0
+ */
+import { Applicative } from 'fp-ts/lib/Applicative'
+import { Predicate, Refinement } from 'fp-ts/lib/function'
+import { HKT, Kind, Kind2, Kind3, URIS, URIS2, URIS3 } from 'fp-ts/lib/HKT'
+import { identity } from 'fp-ts/lib/Identity'
+import { Traversable, Traversable1, Traversable2, Traversable3 } from 'fp-ts/lib/Traversable'
+import { ModifyF } from '.'
+import * as I from './internal'
+
+// -------------------------------------------------------------------------------------
+// model
+// -------------------------------------------------------------------------------------
+
+/**
+ * @category model
+ * @since 2.3.0
+ */
+export interface Traversal<S, A> {
+  readonly modifyF: ModifyF<S, A>
+}
+
+// -------------------------------------------------------------------------------------
+// constructors
+// -------------------------------------------------------------------------------------
+
+/**
+ * Create a `Traversal` from a `Traversable`
+ *
+ * @category constructor
+ * @since 2.3.0
+ */
+export function fromTraversable<T extends URIS3>(T: Traversable3<T>): <R, E, A>() => Traversal<Kind3<T, R, E, A>, A>
+export function fromTraversable<T extends URIS2>(T: Traversable2<T>): <E, A>() => Traversal<Kind2<T, E, A>, A>
+export function fromTraversable<T extends URIS>(T: Traversable1<T>): <A>() => Traversal<Kind<T, A>, A>
+export function fromTraversable<T>(T: Traversable<T>): <A>() => Traversal<HKT<T, A>, A>
+export function fromTraversable<T>(T: Traversable<T>): <A>() => Traversal<HKT<T, A>, A> {
+  return <A>() => ({
+    modifyF: <F>(F: Applicative<F>) => {
+      const traverseF = T.traverse(F)
+      return (f: (a: A) => HKT<F, A>) => (s: HKT<T, A>) => traverseF(s, f)
+    }
+  })
+}
+
+// -------------------------------------------------------------------------------------
+// compositions
+// -------------------------------------------------------------------------------------
+
+/**
+ * Compose a `Traversal` with an `Iso`
+ *
+ * @category compositions
+ * @since 2.3.0
+ */
+export const composeIso = I.traversalComposeIso
+
+/**
+ * Compose a `Traversal` with a `Lens`
+ *
+ * @category compositions
+ * @since 2.3.0
+ */
+export const composeLens = I.traversalComposeLens
+
+/**
+ * Compose a `Traversal` with a `Prism`
+ *
+ * @category compositions
+ * @since 2.3.0
+ */
+export const composePrism = I.traversalComposePrism
+
+/**
+ * Compose a `Traversal` with a `Optional`
+ *
+ * @category compositions
+ * @since 2.3.0
+ */
+export const composeOptional = I.traversalComposeOptional
+
+/**
+ * Compose a `Traversal` with a `Traversal`
+ *
+ * @category compositions
+ * @since 2.3.0
+ */
+export const compose = I.traversalComposeTraversal
+
+// -------------------------------------------------------------------------------------
+// combinators
+// -------------------------------------------------------------------------------------
+
+/**
+ * @category combinators
+ * @since 2.3.0
+ */
+export const modify = <A>(f: (a: A) => A) => <S>(sa: Traversal<S, A>): ((s: S) => S) => {
+  return sa.modifyF(identity)(f)
+}
+
+/**
+ * @category combinators
+ * @since 2.3.0
+ */
+export const set = <A>(a: A): (<S>(sa: Traversal<S, A>) => (s: S) => S) => {
+  return modify(() => a)
+}
+
+/**
+ * @category combinators
+ * @since 2.3.0
+ */
+export function filter<A, B extends A>(refinement: Refinement<A, B>): <S>(traversal: Traversal<S, A>) => Traversal<S, B>
+export function filter<A>(predicate: Predicate<A>): <S>(traversal: Traversal<S, A>) => Traversal<S, A>
+export function filter<A>(predicate: Predicate<A>): <S>(traversal: Traversal<S, A>) => Traversal<S, A> {
+  return composePrism(I.prismFromPredicate(predicate))
+}
+
+/**
+ * Return a `Traversal` from a `Traversal` and a prop
+ *
+ * @category combinators
+ * @since 2.3.0
+ */
+export const prop = I.traversalProp
+
+/**
+ * Return a `Traversal` from a `Traversal` and a list of props
+ *
+ * @category combinators
+ * @since 2.3.0
+ */
+export const props = I.traversalProps