+++ test/2.2/Index.ts
@@ -1,11 +1,11 @@
-import { indexArray } from '../src/Index/Array'
-import { indexNonEmptyArray } from '../src/Index/NonEmptyArray'
-import { indexRecord } from '../src/Index/Record'
+import { indexArray } from '../../src/Index/Array'
+import { indexNonEmptyArray } from '../../src/Index/NonEmptyArray'
+import { indexRecord } from '../../src/Index/Record'
 import * as assert from 'assert'
 import { some, none } from 'fp-ts/lib/Option'
 import * as R from 'fp-ts/lib/Record'
 import { cons } from 'fp-ts/lib/NonEmptyArray'
-import { Iso } from '../src'
+import { Iso } from '../../src'
 
 describe('Index', () => {
   describe('indexRecord', () => {