+++ test/2.2/Setter.ts
@@ -1,4 +1,4 @@
-import { Setter } from '../src'
+import { Setter } from '../../src'
 import * as assert from 'assert'
 
 type Point = { x: number; y: number }